refactor(dccpeep): extract line program infrastructure
Move shared line storage, user-asm metadata, exact matching, comment stripping, tagged replacement, insertion/deletion, and file I/O into peep_lines.c behind dccpeep_internal.h.

Update m.sh, mmacos.sh, m.bat, scripts/build-dcc.ps1, and a focused src/dccpeep/CMakeLists.txt to compile every dccpeep translation unit as portable C11.

Validation: direct and full PowerShell builds passed; 316 raw assembly inputs produced byte-identical optimized output against the post-pass_once monolith; no performance baselines changed.
