+++ mmacos.sh
@@ -1,7 +1,7 @@
 rm -f dcc dccpeep dccrtlstrip dccmake m80c
 
 sh src/dcc/build-dcc.sh
-clang -std=c89 -Wall -Wextra -O2 -g -o dccpeep src/dccpeep/dccpeep.c
+clang -std=c11 -Wall -Wextra -O2 -g -I src/dccpeep -o dccpeep src/dccpeep/*.c
 clang -std=c89 -Wall -Wextra -O2 -g -o dccrtlstrip src/dccrtlstrip/dccrtlstrip.c
 clang -std=c89 -Wall -Wextra -O2 -g -o dccmake src/dccmake/dccmake.c
 clang -std=c89 -Wall -Wextra -O2 -g -o m80c src/m80c/m80c.c
